Moved editing host config plugin to sitecore-jss-nextjs-editing-host. Fixed linting errors.
