+++ packages/sitecore-jss-nextjs-editing-host/package.json
@@ -39,6 +39,7 @@
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "ts-node": "^9.0.0",
+    "tslint-react": "^4.0.0",
     "typescript": "^4.0.2"
   },
   "peerDependencies": {
