+++ packages/sitecore-jss-nextjs-editing-host/src/index.ts
@@ -2,3 +2,5 @@ export {
   EditingServerOptions,
   startEditingServer
 } from './editing-server';
+
+export { config } from './config';
